Made prefetching a configurable option; documented settings.
