+++ .gitignore
@@ -3,3 +3,4 @@ target
 data
 data2
 *.iml
+/cache/
