--- src/main/java/de/corelogics/mediaview/service/downloader/ClipMetadata.java
@@ -1,77 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Mediatheken DLNA Bridge Authors.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package de.corelogics.mediaview.service.downloader;
-
-import java.util.BitSet;
-import java.util.StringJoiner;
-
-class ClipMetadata {
-    private String contentType;
-    private long size;
-    private BitSet bitSet;
-    private int numberOfChunks;
-
-    public String getContentType() {
-        return contentType;
-    }
-
-    public void setContentType(String contentType) {
-        this.contentType = contentType;
-    }
-
-    public long getSize() {
-        return size;
-    }
-
-    public void setSize(long size) {
-        this.size = size;
-    }
-
-    public void setBitSet(BitSet bitSet) {
-        this.bitSet = bitSet;
-    }
-
-    public BitSet getBitSet() {
-        return bitSet;
-    }
-
-    public void setNumberOfChunks(int numberOfChunks) {
-        this.numberOfChunks = numberOfChunks;
-    }
-
-    public int getNumberOfChunks() {
-        return numberOfChunks;
-    }
-
-    @Override
-    public String toString() {
-        return new StringJoiner(", ", ClipMetadata.class.getSimpleName() + "[", "]")
-                .add("contentType='" + contentType + "'")
-                .add("size=" + size)
-                .add("bitSet=" + bitSet)
-                .add("numberOfChunks=" + numberOfChunks)
-                .toString();
-    }
-}
