+++ src/main/java/de/corelogics/mediaview/service/proxy/downloader/ClipChunk.java
@@ -22,14 +22,33 @@
  * SOFTWARE.
  */
 
-package de.corelogics.mediaview.service.downloader;
+package de.corelogics.mediaview.service.proxy.downloader;
 
-public class UpstreamReadFailedException extends Exception {
-    public UpstreamReadFailedException(String message) {
-        super(message);
+class ClipChunk {
+    private final int chunkNumber;
+    private final long from;
+    private final long to;
+
+    public ClipChunk(int chunkNumber, long from, long to) {
+        this.chunkNumber = chunkNumber;
+        this.from = from;
+        this.to = to;
     }
 
-    public UpstreamReadFailedException(String message, Throwable cause) {
-        super(message, cause);
+    public long getFrom() {
+        return from;
+    }
+
+    public long getTo() {
+        return to;
+    }
+
+    public int getChunkNumber() {
+        return chunkNumber;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("{#%d: %d-%d}", chunkNumber, from, to);
     }
 }
