+++ src/main/java/de/corelogics/mediaview/service/proxy/downloader/ClipDownloaderHolder.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package de.corelogics.mediaview.service.downloader;
+package de.corelogics.mediaview.service.proxy.downloader;
 
 import java.io.EOFException;
 import java.io.FilterInputStream;
