+++ src/main/java/de/corelogics/mediaview/service/proxy/downloader/OpenedStream.java
@@ -22,14 +22,42 @@
  * SOFTWARE.
  */
 
-package de.corelogics.mediaview.service.downloader;
+package de.corelogics.mediaview.service.proxy.downloader;
 
-public class UpstreamNotFoundException extends Exception {
-    public UpstreamNotFoundException(String url, int statusCode) {
-        super(String.format("Upstream not found, status %d at [%s]", statusCode, url));
+import org.h2.util.IOUtils;
+
+import java.io.Closeable;
+import java.io.InputStream;
+
+public class OpenedStream implements Closeable {
+    private final String contentType;
+    private final long maxSize;
+    private InputStream stream;
+
+    public OpenedStream(String contentType, long maxSize, InputStream stream) {
+        this.contentType = contentType;
+        this.maxSize = maxSize;
+        this.stream = stream;
     }
 
-    public UpstreamNotFoundException(String url, int statusCode, Throwable cause) {
-        super(String.format("Upstream not found, status %d at [%s]", statusCode, url), cause);
+    public String getContentType() {
+        return contentType;
+    }
+
+    public long getMaxSize() {
+        return maxSize;
+    }
+
+    public InputStream getStream() {
+        return stream;
+    }
+
+    public void setStream(InputStream stream) {
+        this.stream = stream;
+    }
+
+    @Override
+    public void close() {
+        IOUtils.closeSilently(stream);
     }
 }
