+++ src/main/resources/application.properties
@@ -25,5 +25,10 @@ DISPLAY_NAME=Mediatheken
 MEDIATHEKVIEW_LIST_BASEURL=https://res.mediathekview.de
 DATABASE_LOCATION=./data/clipdb
 UPDATEINTERVAL_FULL_HOURS=24
+ENABLE_PREFETCHING=false
+PUBLIC_HTTP_PORT=8080
 CACHE_DIRECTORY=./cache
 CACHE_SIZE_GB=10
+CACHE_DOWNLODERS_PER_VIDEO=2
+CACHE_MAX_PARALLEL_DOWNLOADS=4
+PUBLIC_BASE_URL=http://localhost:8080
