+++ src/main/resources/log4j2.xml
@@ -25,7 +25,7 @@
 
 <Configuration xmlns="http://logging.apache.org/log4j/2.0/config"
                      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                     xsi:schemaLocation="http://jakarta.apache.org/log4j/ https://raw.githubusercontent.com/apache/logging-log4j2/master/log4j-core/src/main/resources/Log4j-config.xsd">
+>
     <Appenders>
         <Console name="Console" target="SYSTEM_OUT">
             <PatternLayout pattern="%d{${LOG_DATEFORMAT_PATTERN:-yyyy-MM-dd HH:mm:ss.SSS}} %highlight{${LOG_LEVEL_PATTERN:-%5p}}{FATAL=red blink, ERROR=red, WARN=yellow bold, INFO=green, DEBUG=green bold, TRACE=blue} [%15.15t] %style{%-40.40C{1.}}{cyan} : %m%n"/>
@@ -33,7 +33,7 @@
     </Appenders>
     <Loggers>
         <Logger name="org.fourthline.cling.protocol.RetrieveRemoteDescriptors" level="error"/>
-        <Logger name="de.corelogics.mediaview" level="debug"/>
+        <Logger name="de.corelogics.mediaview" level="info"/>
         <Root level="warn">
             <AppenderRef ref="Console"/>
         </Root>
